requirements.txt
flask
flask-cors
flask-jwt-extended
python-dotenv
supabase
ultralytics
torch
opencv-python